+++ changelog.md
@@ -14,6 +14,8 @@ Codion Change Log
 - EntitySearchModel.PERSISTENCE_EVENTS renamed PERSISTENCE_AWARE, Builder.persistenceEvents() renamed persistenceAware().
 - PersistenceEvents.events() renamed persistenceEvents() for static import.
 - DefaultEntityEditModel.DefaultSettings minor refactoring.
+- EntityEditModel.applyEdit() entities Collection parameter replaced with a single entity.
+- DefaultEntityEditor now uses Objects.deepEquals() when notifying.
 ### is.codion.swing.framework.model
 - EntityComboBoxModel.PERSISTENCE_EVENTS renamed PERSISTENCE_AWARE, Builder.persistenceEvents() renamed persistenceAware().
 ### is.codion.swing.framework.ui